Fixed pipes

+++ src/pipe/create_pipe.c
@@ -1,9 +1,6 @@
-#include <los/syscall.h>
 #include <los/pipe.h>
+#include <los/syscall.h>
 
-
-
-isize create_pipe(usize* prd, usize* pwd){
-    return system_call2(CREATE_PIPE_SYSCALL, prd, pwd);
-
+isize create_pipe(usize* prd, usize* pwd) {
+    return system_call2(CREATE_PIPE_SYSCALL, (usize)prd, (usize)pwd);
 }
