+++ src/pipe/read_pipe.c
@@ -1,7 +1,6 @@
-#include <los/syscall.h>
 #include <los/pipe.h>
+#include <los/syscall.h>
 
-
-isize read_pipe(usize prd, void* buffer, usize buffer_len){
-    return system_call3(READ_PIPE_SYSCALL, prd, buffer, buffer_len);
-} 
+isize read_pipe(usize prd, void* buffer, usize buffer_len) {
+    return system_call3(READ_PIPE_SYSCALL, prd, (usize)buffer, buffer_len);
+}
