+++ src/pipe/write_pipe.c
@@ -1,8 +1,6 @@
-#include <los/syscall.h>
 #include <los/pipe.h>
+#include <los/syscall.h>
 
-
-
-isize write_pipe(usize pwd, void* buffer, usize buffer_len){
-    return system_call3(WRITE_PIPE_SYSCALL, pwd, buffer, buffer_len);
+isize write_pipe(usize pwd, void* buffer, usize buffer_len) {
+    return system_call3(WRITE_PIPE_SYSCALL, pwd, (usize)buffer, buffer_len);
 }
